feat: 카테고리 조회 API 기능 구현

+++ src/main/java/com/poortorich/category/repository/CategoryRepository.java
@@ -4,9 +4,7 @@ import com.poortorich.category.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, BigInteger> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 
 }
